fix remaining spells for midnight
